Fix missing format arg in custom exception; update all unit tests

+++ src/main/java/com/meutley/studioadmintoolkit/invoice/InvoiceServiceImpl.java
@@ -38,6 +38,10 @@ public class InvoiceServiceImpl implements InvoiceService {
 
     @Override
     public void deleteLineItem(int invoiceId, int lineItemId) {
+        InvoiceLineItem lineItem = this.invoiceLineItemRepository.getOneForInvoice(invoiceId, lineItemId);
+        if (lineItem == null) {
+            throw new EntityNotFoundException(lineItemId);
+        }
         this.invoiceLineItemRepository.deleteById(lineItemId);
     }
